requirements.txt
imutils~=0.5.4
opencv-python~=4.7.0.72
Flask~=2.2.3
scikit-image~=0.20.0
Pillow~=9.4.0